AE-97 chore: move api bridge to old apps-engine folder
